move the robot in square in loop
This program moves the robot the in 1m square , in continious loop. This will help to understand the odometry error in the program. If the robot comes back to orignal position (from where it started) then it means no error. If it does not come then it means it has some error in odometry.